requirements.txt
# Core web framework
fastapi
uvicorn
jinja2

# Vector search & embeddings
sentence-transformers
faiss-cpu
numpy
torch

# Retrieval & document processing
langchain
langchain-community
pypdf
tiktoken
ocrmypdf

# Utilities & integrations
requests
pynvml
pyyaml
matplotlib
pandas
nltk
rouge-score
